rename: Product, ProductList를 Product폴더로 이동
